refactor(cli); Pull out config
